Housekeeping: Move types from viewmodel to entities
Housekeeping: Moved entities from `view-model/page-did.d.ts` to entities

Housekeeping: moved widget types to a new file in entities

Housekeeping: Moved infrastructure types for charts
Moved the infrastructure types for charts in widgets to the
component-library, mirroring the structure of the StreamedTable
`types.d.ts` file.

Housekeeping: Moved RSE Types to Entity from viewmodels.

Fix imports
